additional steps

+++ LibraryManagement/src/main/java/learn/domain/BorrowService.java
@@ -2,6 +2,7 @@ package learn.domain;
 
 import learn.data.BorrowRepository;
 import learn.models.Borrow;
+import learn.models.User;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -10,26 +11,102 @@ import java.util.Optional;
 @Service
 public class BorrowService {
 
-    private BorrowRepository borrowRepository;
+    private final BorrowRepository borrowRepository;
 
     public BorrowService(BorrowRepository borrowRepository) {
         this.borrowRepository = borrowRepository;
     }
 
-    public Optional<Borrow> findById(Long id){
+    public Optional<Borrow> findById(java.lang.Long id){
         return borrowRepository.findById(id);
+    }
+
+    public List<Borrow> findByUser(User userId){
+        return borrowRepository.findByUser(userId);
+    }
+
+    public List<Borrow> findByBook(User bookId){
+        return borrowRepository.findByBook(bookId);
     }
 
     public List<Borrow> findAll(){
         return borrowRepository.findAll();
     }
 
-    public Borrow addBorrow(Borrow borrow){
-        return borrowRepository.save(borrow);
+    public Result<Borrow> addBorrow(Borrow borrow){
+        Result<Borrow> result = validate(borrow);
+
+        if(borrow.getId() != null){
+            result.addMessage("Borrow Id cannot be set for add operation", ResultType.INVALID);
+        }
+
+        if(!result.isSuccess()){
+            return result;
+        }
+
+        result.setPayload(borrowRepository.save(borrow));
+        return result;
     }
 
-    public void deleteById(Long id){
-        borrowRepository.deleteById(id);
+    public Result<Borrow> updateBorrow(Borrow updatedBorrow) {
+        Result<Borrow> result = validate(updatedBorrow);
+
+        if (updatedBorrow.getId() == null) {
+            result.addMessage("BorrowId is required", ResultType.INVALID);
+        }
+
+        if (!result.isSuccess()) {
+            return result;
+        }
+
+        java.lang.Long id = updatedBorrow.getId();
+        Optional<Borrow> existingBorrowOpt = borrowRepository.findById(id);
+        if (existingBorrowOpt.isPresent()) {
+            Borrow existingBorrow = existingBorrowOpt.get();
+            existingBorrow.setUser(updatedBorrow.getUser());
+            existingBorrow.setBook(updatedBorrow.getBook());
+            existingBorrow.setBorrowDate(updatedBorrow.getBorrowDate());
+            existingBorrow.setReturnDate(updatedBorrow.getReturnDate());
+            result.setPayload(borrowRepository.save(existingBorrow));
+        } else {
+            result.addMessage("Borrow ID " + id + " not found.", ResultType.NOT_FOUND);
+        }
+
+        return result;
+    }
+
+    public Result<Borrow> deleteById(java.lang.Long id){
+        Result<Borrow> result = new Result<>();
+        if (borrowRepository.existsById(id)) {
+            borrowRepository.deleteById(id);
+        } else {
+            result.addMessage("Book ID " + id + " not found.", ResultType.NOT_FOUND);
+        }
+        return result;
+    }
+
+    private Result<Borrow> validate(Borrow borrow) {
+        Result<Borrow> result = new Result<>();
+        if (borrow == null) {
+            result.addMessage("Borrow cannot be null", ResultType.INVALID);
+            return result;
+        }
+
+        if (borrow.getBorrowDate() == null) {
+            result.addMessage("DateOut is required", ResultType.INVALID);
+        }
+
+        if (borrow.getReturnDate() == null) {
+            result.addMessage("Return Date is required", ResultType.INVALID);
+        }
+
+        if (borrow.getReturnDate() != null && borrow.getBorrowDate() != null) {
+            if (borrow.getReturnDate().isBefore(borrow.getBorrowDate())) {
+                result.addMessage("Return Date must be after Borrow Date", ResultType.INVALID);
+            }
+        }
+
+        return result;
     }
 
 }
